Implemented hiding paddle when ball is not near
Added extra commandline option `-d` to allow user to set a
distance in pixels. When the horizontal distance between the paddle
and the ball is more than the given distance the paddle will not
be drawn. By default the distance is set to the width of the window,
effectively always drawing the paddle.